+++ src/storage/page/b_plus_tree_leaf_page.cpp
@@ -101,6 +101,6 @@ auto B_PLUS_TREE_LEAF_PAGE_TYPE::UpperBoundByFirst(const KeyType &key, const Key
   return l;
 }
 // template class BPlusTreeLeafPage<key_t, page_id_t, Comparator>;
-template class BPlusTreeLeafPage<pair<size_t, int>, int,Comparator<size_t ,int,int>>;
+template class BPlusTreeLeafPage<pair<uint64_t, int>, int, Comparator<uint64_t, int, int>>;
 template class BPlusTreeLeafPage<pair<String<65>, int>, int, Comparator<String<65>, int, int>>;
 }  // namespace CrazyDave
